tests/wpt/tests/html/canvas/offscreen/path-objects/2d.path.arcTo.shape.start.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.path.arcTo.shape.start</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.path.arcTo.shape.start</h1>
<p class="desc">arcTo() draws a straight line from P0 to P1</p>


<script>
var t = async_test("arcTo() draws a straight line from P0 to P1");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.fillStyle = '#f00';
  ctx.fillRect(0, 0, 100, 50);
  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 50;
  ctx.beginPath();
  ctx.moveTo(0, 25);
  ctx.arcTo(200, 25, 200, 50, 10);
  ctx.stroke();

  _assertPixel(canvas, 1,1, 0,255,0,255);
  _assertPixel(canvas, 1,48, 0,255,0,255);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  _assertPixel(canvas, 98,1, 0,255,0,255);
  _assertPixel(canvas, 98,48, 0,255,0,255);
  t.done();

});
</script>
